requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.1
pymongo==4.10.1
python-dotenv==1.0.1
